Novos exercicios.

+++ exercicios/ex008.py
@@ -1,3 +1,3 @@
-m = int(input('Digite o valor em metro: '))
+m = float(input('Digite o valor em metro: '))
 
 print('O valor em centímetros é {}.\nO valor em milimetros é {}.'.format(m*100,m*1000))
